+++ packages/solid/src/components/field/field.stories.tsx
@@ -7,5 +7,6 @@ const meta: Meta = {
 export default meta
 
 export { Input } from './examples/input'
+export { RequiredIndicator } from './examples/required-indicator'
 export { Select } from './examples/select'
 export { Textarea } from './examples/textarea'