+++ packages/solid/src/components/field/index.tsx
@@ -11,6 +11,11 @@ export {
 } from './field-helper-text'
 export { FieldInput, type FieldInputBaseProps, type FieldInputProps } from './field-input'
 export { FieldLabel, type FieldLabelBaseProps, type FieldLabelProps } from './field-label'
+export {
+  FieldRequiredIndicator,
+  type FieldRequiredIndicatorBaseProps,
+  type FieldRequiredIndicatorProps,
+} from './field-required-indicator'
 export { FieldRoot, type FieldRootBaseProps, type FieldRootProps } from './field-root'
 export {
   FieldRootProvider,